Make UserPermissions Admin panel more modular and export components
