+++ packages/admin/cms-admin/src/userPermissions/UserGrid.tsx
@@ -5,24 +5,29 @@ import {
     GridFilterButton,
     muiGridFilterToGql,
     muiGridSortToGql,
-    StackSwitchApiContext,
+    ToolbarActions,
+    ToolbarFillSpace,
     ToolbarItem,
     useDataGridRemote,
     usePersistentColumnState,
 } from "@comet/admin";
-import { Edit } from "@comet/admin-icons";
-import { IconButton, Typography } from "@mui/material";
+import { Typography } from "@mui/material";
 import { styled } from "@mui/material/styles";
-import { DataGrid, GridToolbarQuickFilter } from "@mui/x-data-grid";
-import { useContext } from "react";
+import { DataGrid, GridRenderCellParams, GridToolbarQuickFilter } from "@mui/x-data-grid";
 import { useIntl } from "react-intl";
 
 import { GQLUserForGridFragment, GQLUserGridQuery, GQLUserGridQueryVariables } from "./UserGrid.generated";
 
-export const UserGrid = () => {
+type Props = {
+    toolbarAction?: React.ReactNode;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    rowAction?: (params: GridRenderCellParams<any, GQLUserForGridFragment, any>) => React.ReactNode;
+    actionsColumnWidth?: number;
+};
+
+export const UserGrid = ({ toolbarAction, rowAction, actionsColumnWidth = 52 }: Props) => {
     const dataGridProps = { ...useDataGridRemote(), ...usePersistentColumnState("UserGrid") };
     const intl = useIntl();
-    const stackApi = useContext(StackSwitchApiContext);
 
     const columns: GridColDef<GQLUserForGridFragment>[] = [
         {
@@ -46,17 +51,15 @@ export const UserGrid = () => {
             field: "actions",
             headerName: "",
             sortable: false,
-            pinnable: false,
             filterable: false,
-            renderCell: (params) => (
-                <IconButton
-                    onClick={() => {
-                        stackApi.activatePage("edit", params.id.toString());
-                    }}
-                >
-                    <Edit color="primary" />
-                </IconButton>
-            ),
+            type: "actions",
+            align: "right",
+            pinned: "right",
+            width: actionsColumnWidth,
+            disableExport: true,
+            renderCell: (params) => {
+                return <> {rowAction && rowAction(params)}</>;
+            },
         },
     ];
 
@@ -104,8 +107,13 @@ export const UserGrid = () => {
                         <ToolbarItem>
                             <GridFilterButton />
                         </ToolbarItem>
+                        <ToolbarFillSpace />
+                        {toolbarAction && <ToolbarActions>{toolbarAction}</ToolbarActions>}
                     </DataGridToolbar>
                 ),
+            }}
+            componentsProps={{
+                toolbar: { toolbarAction },
             }}
         />
     );
